server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
setuptools>=65.0.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
supabase==2.5.1
requests==2.32.3
huggingface_hub==0.29.0
moviepy==1.0.3
Pillow==10.4.0
imageio-ffmpeg==0.4.9
gradio_client==1.5.2